Add multipath logic
Add a toggle to remove 'N' from the iostat options, which forces device
mapper disks to show as their dm- name.

Add a toggle to skip physical disks part of a multipath setup. This
option causes DM_MULTIPATH_DEVICE_PATH to be evaluated. If it has a
value of '1', the disk is a physical disk part of a multipath disk
configuration, and is skipped.

This allows us to evaluate dm- devices and get combined/correct stats
for round-robin multipath disks.

The code to lookup a specific/persistent udev name attribute was moved
and is now evaluated once per disk, instead of once per iostat column.